fix ui bugs
