--- src/main/java/seedu/address/ui/CommandBox.java
@@ -1,96 +1,0 @@
-package seedu.address.ui;
-
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.layout.Region;
-
-import seedu.address.commons.exceptions.AlfredModelHistoryException;
-import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.parser.exceptions.ParseException;
-
-/**
- * The UI component that is responsible for receiving user command inputs.
- */
-public class CommandBox extends UiPart<Region> {
-
-    public static final String ERROR_STYLE_CLASS = "error";
-    private static final String FXML = "CommandBox.fxml";
-
-    private final CommandExecutor commandExecutor;
-
-    @FXML
-    private AutoCompleteTextField commandTextField;
-
-    @FXML
-    private Button sendButton;
-
-
-    public CommandBox(CommandExecutor commandExecutor) {
-        super(FXML);
-        this.commandExecutor = commandExecutor;
-        // calls #setStyleToDefault() whenever there is a change to the text of the command box.
-        commandTextField.textProperty().addListener((unused1, unused2, unused3) -> setStyleToDefault());
-    }
-
-    public void setFocusTextField() {
-        this.commandTextField.requestFocus();
-    }
-
-    /**
-     * Sets the text to be displayed in the TextField {@code commandTextField}.
-     *
-     * @param text
-     */
-    public void setTextField(String text) {
-        this.commandTextField.setText(text);
-    }
-
-    /**
-     * Handles the Enter button pressed event.
-     */
-    @FXML
-    private void handleCommandEntered() {
-        try {
-            commandExecutor.execute(commandTextField.getText());
-            commandTextField.setText("");
-        } catch (CommandException | ParseException | AlfredModelHistoryException e) {
-            setStyleToIndicateCommandFailure();
-        }
-    }
-
-    /**
-     * Sets the command box style to use the default style.
-     */
-    private void setStyleToDefault() {
-        commandTextField.getStyleClass().remove(ERROR_STYLE_CLASS);
-    }
-
-    /**
-     * Sets the command box style to indicate a failed command.
-     */
-    private void setStyleToIndicateCommandFailure() {
-        ObservableList<String> styleClass = commandTextField.getStyleClass();
-
-        if (styleClass.contains(ERROR_STYLE_CLASS)) {
-            return;
-        }
-
-        styleClass.add(ERROR_STYLE_CLASS);
-    }
-
-    /**
-     * Represents a function that can execute commands.
-     */
-    @FunctionalInterface
-    public interface CommandExecutor {
-        /**
-         * Executes the command and returns the result.
-         *
-         * @see seedu.address.logic.Logic#execute(String)
-         */
-        CommandResult execute(String commandText) throws CommandException, ParseException, AlfredModelHistoryException;
-    }
-
-}
