+++ src/main/java/seedu/address/ui/EntityCard.java
@@ -100,7 +100,6 @@ public class EntityCard extends UiPart<Region> {
      */
     private void generateParticipantCard(Entity entity) {
         this.idIcon.setImage(getImage(STUDENT_ICON));
-        logger.info("The participant icon has been changed to: " + idIcon);
         Participant participant = (Participant) entity;
         labels.getChildren().add(new Label(participant.getPhone().value));
         labels.getChildren().add(new Label(participant.getEmail().value));
@@ -155,10 +154,8 @@ public class EntityCard extends UiPart<Region> {
         locationLabel.setStyle("-fx-background-color: #26428b");
         labels.getChildren().add(locationLabel);
         List<Participant> participants = team.getParticipants();
-        logger.info("Number of Members in team: " + participants.size());
         membersPane.getChildren().add(new Label("Members: "));
 
-        logger.info("Size of membersPane before adding Participants: " + membersPane.getChildren().size());
         String members = participants.stream().map(p -> p.getName().toString()).collect(Collectors.joining(" | "));
         membersPane.getChildren().add(new Label(members));
         this.type = PrefixType.T;
@@ -197,8 +194,6 @@ public class EntityCard extends UiPart<Region> {
     }
 
     private Image getImage(String imagePath) {
-        logger.info("File path of image is: " + imagePath);
-        logger.info(this.getClass().getResourceAsStream(imagePath).toString());
         return new Image(Objects.requireNonNull(this.getClass().getResourceAsStream(imagePath)));
     }
