+++ src/main/java/seedu/address/ui/listpanel/EntityListPanel.java
@@ -35,8 +35,6 @@ public class EntityListPanel extends UiPart<Region> {
     public EntityListPanel(ObservableList<? extends Entity> entityList) {
         super(FXML);
         entityList.forEach(item -> listView.getItems().add(item));
-        logger.info("Size of EntityList is: " + listView.getItems().size());
-        logger.info("Size of EntityListView is: " + listView.getItems().size());
         if (!entityList.isEmpty()) {
             Entity firstItem = entityList.get(0);
             if (firstItem instanceof Participant) {
@@ -49,7 +47,7 @@ public class EntityListPanel extends UiPart<Region> {
                 prefix = PrefixType.M;
                 listView.setCellFactory(listView -> new MentorListViewCell());
             }
-            logger.info("EntityListView has prefix type: " + this.prefix);
+
         }
     }
 
@@ -65,10 +63,8 @@ public class EntityListPanel extends UiPart<Region> {
             if (isEmpty || curr == null) {
                 setGraphic(null);
                 setText(null);
-                logger.info("Item does not exist");
             } else {
                 setGraphic(new EntityCard(curr, getIndex() + 1).getRoot());
-                logger.info("Graphic is set to EntityCard of Mentor type");
             }
         }
     }
@@ -85,10 +81,8 @@ public class EntityListPanel extends UiPart<Region> {
             if (isEmpty || curr == null) {
                 setGraphic(null);
                 setText(null);
-                logger.info("Item does not exist");
             } else {
                 setGraphic(new EntityCard(curr, getIndex() + 1).getRoot());
-                logger.info("Graphic is set to EntityCard of Participant type");
             }
         }
     }
@@ -105,10 +99,8 @@ public class EntityListPanel extends UiPart<Region> {
             if (isEmpty || curr == null) {
                 setGraphic(null);
                 setText(null);
-                logger.info("Item does not exist");
             } else {
                 setGraphic(new EntityCard(curr, getIndex() + 1).getRoot());
-                logger.info("Graphic is set to EntityCard of Team type");
             }
         }
     }
