+++ src/main/resources/view/AutoCompleTextField.css
@@ -1,0 +1,25 @@
+.context-menu {
+
+    -fx-background-color: #e3f9ff;
+    -fx-background-radius: 0 6 6 6;
+
+}
+
+.menu-item {
+ -fx-font-size: 15;
+}
+.menu-item:focused {
+     -fx-background: -fx-accent;
+     -fx-background-color: #d6e5ff;
+     -fx-text-fill: #ffffff;
+}
+
+.jfx-text-field {
+   -fx-pref-height: 60;
+   -fx-prompt-text-fill: #cef9fb;
+   -fx-text-fill: #cef9fb;
+   -jfx-focus-color: #cef9fb;
+   -fx-font-family: Arial;
+   -fx-font-size: 16;
+}
+
