--- src/main/resources/view/CommandBox.fxml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-<?import seedu.address.ui.*?>
-
-<HBox style="-fx-background-color: #292734;" styleClass="stack-pane" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1">
-  <AutoCompleteTextField fx:id="commandTextField" onAction="#handleCommandEntered" prefHeight="60" promptText="What can I do for you?" style="-fx-prompt-text-fill: #cef9fb; -fx-text-fill: #cef9fb; -jfx-focus-color: #cef9fb;" stylesheets="@ContextMenu.css" HBox.hgrow="ALWAYS">
-      <font>
-         <Font name="Arial" size="16.0" />
-      </font></AutoCompleteTextField>
-</HBox>
